src/requirements.txt
numpy==2.4.2
typing_extensions==4.15.0
eclipse-zenoh==0.4.0
colorlog==6.10.1
